+++ Sources/QLoop/Iterating/QLoopIteratorContinueNil.swift
@@ -4,8 +4,6 @@ public final class QLoopIteratorContinueNil: QLoopIterating {
     public init() {
     }
 
-    public func reset() { /**/ }
-
     @discardableResult
     public func iterate(_ loop: QLoopIterable) -> Bool {
         guard (loop.discontinue == false) else { return false }